requirement.txt
Django>=4.0.0,<5.0.0
djangorestframework>=3.13.0,<4.0.0
django-filter==23.3  
drf-yasg==1.21.7    